feat(code): Add `select_proposer` method to `Context` trait
This allows app-developers to customize the proposer selection algorithm.
